pyproject.toml
[project]
name = "gemini-knowledge-mcp"
version = "1.0.0"
description = "MCP Server for Gemini-powered cryptocurrency knowledge"
dependencies = [
    "google-genai>=0.5.0",
    "python-dotenv>=1.0.0",
    "mcp>=0.5.0"
]

[project.scripts]
gemini-knowledge-server = "mcp_server:main"